pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zoho-crm-mcp-server"
version = "0.1.0"
description = "MCP server for Zoho CRM API integration"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "asklokesh", email = "asklokesh@example.com" }
]
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0"
]

[project.scripts]
zoho-crm-mcp = "zoho_crm_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/zoho_crm_mcp"]
